requirements.txt
requests==2.32.3
pyyaml==6.0.2
tomli==2.2.1
python-dotenv==1.0.1